+++ lib/grape/validations/regexp.rb
@@ -4,8 +4,7 @@ module Grape
     class RegexpValidator < SingleOptionValidator
       def validate_param!(attr_name, params)
         if params[attr_name] && !( params[attr_name].to_s =~ @option )
-          raise Grape::Exceptions::Validation, :status => 400,
-            :param => @scope.full_name(attr_name), :message_key => :regexp
+          raise Grape::Exceptions::Validation, :param => @scope.full_name(attr_name), :message_key => :regexp
         end
       end
     end